Don't build unnecessary sources. Remove tptimer.cpp from DisplaySettings. Remove tptimer.cpp from FrameRate. Remove powerstate.cpp from FrontPanel. Remove unused tptimer.h from HdmiCec. Remove tptimer.cpp from HdmiCecSink. Remove cSettings.cpp and powerstate.cpp from MaintenanceManager. Remove unused powerstate.h from MaintenanceManager. Remove tptimer.cpp from RDKShell. Remove cSettings.cpp and powerstate.cpp from SystemServices. Remove tptimer.cpp from Timer. Remove powerstate.cpp from Warehouse. Remove tptimer.cpp from XCast.
