+++ DisplaySettings/CMakeLists.txt
@@ -23,7 +23,6 @@ find_package(${NAMESPACE}Plugins REQUIRED)
 add_library(${MODULE_NAME} SHARED
         DisplaySettings.cpp
         Module.cpp
-	../helpers/tptimer.cpp
         ../helpers/UtilsSecurityToken.cpp)
 
 set_target_properties(${MODULE_NAME} PROPERTIES
