+++ HdmiCec/HdmiCec.h
@@ -35,7 +35,6 @@
 
 #include "Module.h"
 
-#include "tptimer.h"
 #include <thread>
 #include <mutex>
 #include <condition_variable>
