+++ MaintenanceManager/CMakeLists.txt
@@ -22,9 +22,7 @@ find_package(${NAMESPACE}Plugins REQUIRED)
 
 add_library(${MODULE_NAME} SHARED
         MaintenanceManager.cpp
-        Module.cpp
-        ../helpers/cSettings.cpp
-        ../helpers/powerstate.cpp)
+        Module.cpp)
 
 set_target_properties(${MODULE_NAME} PROPERTIES
         CXX_STANDARD 11
