+++ SystemServices/CMakeLists.txt
@@ -28,8 +28,6 @@ add_library(${MODULE_NAME} SHARED
         SystemServices.cpp
         Module.cpp
         cTimer.cpp
-        ../helpers/cSettings.cpp
-        ../helpers/powerstate.cpp
         thermonitor.cpp
         SystemServicesHelper.cpp
         uploadlogs.cpp
