+++ XCast/CMakeLists.txt
@@ -23,8 +23,7 @@ find_package(${NAMESPACE}Plugins REQUIRED)
 add_library(${MODULE_NAME} SHARED
         XCast.cpp
         Module.cpp
-        RtXcastConnector.cpp
-	../helpers/tptimer.cpp)
+        RtXcastConnector.cpp)
 
 find_package(RFC)
 find_package(IARMBus)
